build: tools: add in roomservice
This is a new implementation of roomservice
use the new github search api to reduce the memory footprint, and make it less processor intensive
allow for unofficial devices to fetch dependencies. just specify your github organization

please note that comments can not go into this file. It must be valid json
Change-Id: I3d7a8d46d3effbb3d2480c136c9b6b6e610b9b2e